Fix up properties page for GStreamer backend. Mostly a matter of calling
* src/totem-properties-page.c: (totem_properties_page_init),
(set_property), (main):
Fix up properties page for GStreamer backend. Mostly a matter of
calling _init() through getting and using the popt tables, and
explicitely re-setting the metadata after playing, because the
GStreamer backend will call got-metadata in an idle loop which
requires a GMainLoop in order to be executed and - yeah! - that's
not the case here.