When we got an error because we're missing the required source or decoder
* src/backend/bacon-video-widget-gst-0.10.c:
(bvw_check_missing_plugins_error):
(bvw_bus_message_cb):
When we got an error because we're missing the required source or
decoder plugins AND the missing-plugin signal was handled, then make
sure we're not processing any further (less precise/generic) error
messages that may also be on the bus.
Convert error to BVW error when an error occurs asynchroneously too.

svn path=/trunk/; revision=3877