Bug 578387 – Remaining API documentation additions
This adds a load of .h files to the ignore list, and documents the remaining
functions which would be useful to plugins.

It makes TOTEM_MAX_RECENT_ITEM_LEN private to totem-menu.c, since that was the
only file which used it.

It also removes a completely unused declaration of bacon_cd_selection_create
from totem-preferences.h.

See the changes to totem-sections.txt for the list of API I've considered
useful to expose (by way of documentation) to plugins.