Bug 164918 — Remember current position when Pause is clicked
Fixes a case where the position of a stream would be remembered if you
paused and closed Totem, then reopened Totem and played the stream to
the end. Next time you opened Totem, the old saved position would be
used.

Also prevents stream positions from being saved if the position is
within a given percentage (of the stream length) of the start of the
file. Re-closes: bgo#164918