Fix saving current position for the wrong playlist item
When skipping to the next item in the playlist, we
used to use totem_get_current_mrl(), which gets the current
item from the playlist.

Except that the playlist thinks we're already at the next item.

Also use different variables to store the startup and the restore
positions, as we don't want to pause newly added items.

https://bugzilla.gnome.org/show_bug.cgi?id=600425