Avoid problems when pausing before preroll
When pausing before preroll, we would check the length of the file
and find that we might we using a live stream, and stopping it, thus
stopping the preroll.

Instead, when pausing when still prerolling, pause, and only when
the preroll has been done, stop the pipeline if pause was requested.