Split setting up and opening browser streams
Wait until we have the browser stream set up before opening stdin
in the browser, which means we might have a chance to get the stream
size.

With recent GStreamer changes, this means that download buffering would
work, and that we'd get download progress and seeking without needing
to wait until the end of the download.

Helps with https://bugzilla.gnome.org/show_bug.cgi?id=608301