Fix constant buffering when playing live streams
When playing Ogg Theora live streams, we should wait until the
buffer reaches 100% before we start the actual playback.
Otherwise the buffer will be consumed as it is downloaded, and
we'll be constantly buffering, which is problematic visually,
and could cause hiccups during playback.

Fixes constant buffering with the streams at:
http://dir.xiph.org/by_format/Ogg_Theora

https://bugzilla.gnome.org/show_bug.cgi?id=617791