Move methods from Totem to TotemObject for use in plugins
g-ir-scanner will assign methods named "totem_foo" to the Totem record, so
they all need to be renamed to "totem_object_foo" so that they're correctly
assigned to the TotemObject class.

This just changes the method declarations and definitions, and adds a #define
for each method so that the old name continues to work in Totem code. This
can be tidied up in a later commit, possibly at the same time as tidying up
Totem's plugin-visible API.