Fix distcheck for the sample-vala plugin
Now that it's been ported to automake 1.11's Vala support, it's more obvious
that the ENABLE_VALA check doesn't belong in the Makefile. The plugin will be
built if Vala support is enabled, and won't be built if it's disabled.

Of course, distcheck is still broken, as it tries to build the sample-vala
plugin. This is correct, since we should distribute the C files for the Vala
plugin, rather than the Vala files, but since Vala support (bindings) haven't
been re-added to Totem, distcheck fails.