Break a reference cycle in TotemPluginsEngine
TotemPluginsEngine held the only reference on our PeasExtensionSet, which
itself held a reference to the TotemPluginsEngine. This ensured that no
plugins were actually deactivated before Totem was closed, leaking memory and
leaving it impossible for the plugins to save state (etc.) before Totem was
closed.

This has been fixed by introducing an explicit
totem_plugins_engine_shut_down() method, which is called right at the start
of Totem's shut down code. It breaks the reference cycle by finalising the
PeasExtensionSet early, before the TotemPluginsEngine is finalised later on.