Offer to open YouTube videos in the browser if their format isn't supported
Now that we have a much slimmer selection of video formats to stream from
YouTube, it's quite possible that a given video won't be available in a
playable format. In this case, we now display a dialogue box to the user when
they try to play the video, asking them if they'd like to watch it in their
web browser instead.