main: Allow finer-grained control on volume
When adjusting volume using the keyboard, I often find that raising or
lowering the volume level by 8% is too drastic a change, and I can't
seem to hit a "comfortable" volume unless I reach for the mouse.

We now bind the Shift-UpArrow and Shift-DownArrow keys to a
volume adjustment of +2% and -2% (respectively), in the spirit of the
fine-grained seek controls (Shift-LeftArrow and Shift-RightArrow).

https://bugzilla.gnome.org/show_bug.cgi?id=641672