playlist: Fix "disappearing" entries
When processing out-of-order playlist entries, make sure to use
their own parsing results.

If we don't do that, we might be using an error result, or an
unhandled result as a success because the earlier item we just parsed
was such a success.

This can be reproduced using mixed types of media over a slow link, eg.
"dual-types" files such as MP4, and "unhandled" types such as AVI over
SMB.