main: Don't poke at web servers for info
When loading videos from some websites, through quvi, we
already have a lot of information about the file, so we
don't really need to go and look for that information
ourselves.

We now store the content-type of the stream in question, and
don't hit the web server to fill in GtkRecent if that's the case.

https://bugzilla.gnome.org/show_bug.cgi?id=644949