main: Fix remote option parsing
When calling a function remotely for the second time, we were
inheriting the options from the first call. Eg. --pause followed by
--play would play and pause instantly the second time.

Note that we cannot do this at the same time as we reset the filenames
because we would end up with no "late" options to handle (such as
session restore).

https://bugzilla.gnome.org/show_bug.cgi?id=669327