backend: Remove non-PulseAudio sink hacks
The goal of the original patch was to make Totem's audio output code
not wake up on a regular basis, as the audio device was still opened.
Now that PulseAudio is used pretty much everywhere we care about, remove
that code, and simplify the logic.

Revert "Only set playbin2 to NULL on errors or after 10 seconds of doing nothing"

This reverts commit 14b0c9e7ca8472fadf59a77bbb43786ee94b72b1.

Conflicts:

	src/backend/bacon-video-widget-gst-0.10.c