main: Don't forget to clear seeking on close
If we had opened a file, got a remembered position to restore,
but never actually got a change in the seekable property, we wouldn't
have clear the seek to position.

This should fix bizarre behaviour if you're skipping between streams
too quickly for GStreamer to see the stream is seekable.

https://bugzilla.gnome.org/show_bug.cgi?id=676434