main: Fix possible assertions when skipping over files
(totem:24342): Gtk-CRITICAL **: gtk_list_store_get_path: assertion `iter->stamp == priv->stamp' failed

When the repeat mode was on, and reaching the end of the playlist,
it was possible to cause a warning as we tried to read past the
last iter in the list store.

We change the semantics of totem_playlist_has_{next,previous}_mrl
to not special case when "repeat" is on.