main: _set_mrl_with_warning() did not work
GStreamer would return the errors asynchronously, so it was impossible
for us to detect errors through bvw_open()'s retval. Remove all the code
that relied on that retval.

We're still left with the same bugs we were before, that some errors are
reported async when we don't want to hear about them.