backend: Fix running with clutter-gst in different prefix
We did not call clutter_gst_init(), so the cluttersink plugin would
not be registered. This meant that we just happened to be able to
get the sink through its (deprecated) stand-alone plugin.

Calling clutter_gst_init() means that we don't need to rely on
libgstclutter.so being available in the GStreamer plugin directory
of the prefix for us to be able to find it.