main: Correctly remove multiple menu items
We can't just iterate over the menu model when removing menu items,
as the item 1 because item 0, when item 0 is removed, etc.

Also make sure that we don't remove associated actions if there is
a target value set, as this is likely to mean that the action is
used by some other menu items.