main: Fixes height < 0 warnings
Show/realizes the main window after the video widget (and removes
the realize of the bvw widget).

Totem used to show the main window, minus the video widget, and then
show the video widget so that any initialisation by the video widget
(like registering plugins) would be done with the window visible.

That's not the case anymore, so we might as well work around those
problems by realizing the video widget in a big enough placeholder.

https://bugzilla.gnome.org/show_bug.cgi?id=703095