backend: Fix pick in the aspect frame
Fixes:
(totem:24473): Clutter-WARNING **: The required ID of 2 does not refer
to an existing actor; this usually implies that the pick() of an actor
is not correctly implemented or that there is an error in the
glReadPixels() implementation of the GL driver.

Let parent do the pick "draw" for the container as per
Example 24. Pick implementation of a container of the clutter
reference manual version 1.8, Section "Implementing a new actor".

https://bugzilla.gnome.org/show_bug.cgi?id=703368