backend: prevent excessive relayouts stopping animations
Every time the elapsed time widget is updating its content, this
triggers a relayout that propagates to the toplevel actor (ie the
stage). Because of the way constraints work, when a relayout is queued
on an actor that is the source of a constraint, it will automatically
queue a relayout on the destination of the constraint too. And because
of the constraints applied between the stage and some of the children,
any relayout triggered by any actor within the scene was relayouting
almost everything, and therefore stopping animations.

This patch removes the usage of constraints in favor of layout
managers that don't have the annoying property of triggering relayouts
on everything.

https://bugzilla.gnome.org/show_bug.cgi?id=700051