main: Add support for starting a playlist paused
Currently, when --pause is specified in the first command line
invocation of Totem, the passed-in media will still start.
This patch saves the initial state of the pause option to a separate
variable, so that when we receive the first playlist changed event, we
can avoid starting up the stream if the option was specified.

https://bugzilla.gnome.org/show_bug.cgi?id=708184