screensaver: Only try to inhibit once
This stops plenty of warnings appearing on the command-line when
gtk_application_inhibit() fails. It can fail for a number of reasons,
but the documentation's only hint is:
If the platform does not support inhibiting or the request failed for
some reason, 0 is returned.

So only try to disable the screensaver once.

https://bugzilla.gnome.org/show_bug.cgi?id=721054