backend: Hide popups when leaving the window
Monitor leave events on the toplevel window, rather than just
on the video widget itself, and only consider it left if it left
to a window that's not in the same hierarchy. This means that
we know that it's left the whole window (and not just the video widget)
and we don't hide the popup when you try to go to the revealed toolbar
in fullscreen.

https://bugzilla.gnome.org/show_bug.cgi?id=727220