grilo: Limit ourselves to running one thumbnailer at a time
Our thumbnailer currently doesn't have support for using hardware
decoding, so it's pretty CPU intensive. On my machine (i7 laptop with an
SSD), one thumbnailer would take about 40% of CPU. This causes both lag
when navigating the videos list, and carries on hampering video playback
if thumbnailing is started before playback.

This is good enough to mark this bug as fixed:
https://bugzilla.gnome.org/show_bug.cgi?id=748370

though a better solution would be to cancel thumbnailing in
totem_grilo_pause(), make sure that
gnome_desktop_thumbnail_path_for_uri() is cancellable, and try to
distinguish in totem-grilo.c between cancellation because the player is
getting shut down, and cancellation because we want all the resources to
go to the player.