plugins: Don't save builtin plugins to GSettings
Builtin plugins are not meant to be disabled by users and are thus not
part of user settings.

Do this by filtering out builtin plugins before writing and inserting
them when reading from the settings key.

This also stops Totem from writing to gsettings on startup. The list of
builtin plugins was not in sync with the default value of the
active-plugins gsettings key.

https://bugzilla.gnome.org/show_bug.cgi?id=749722