backend: embed the fullscreen top header bar in the video widget
On wayland the video has its own subsurface. It is directly painted by the
compositor on top of Totem's main window. As a result on wayland the top
header bar is not visible in fullscreen mode.

To solve this issue, we can embed the top header bar in the video widget.
This solves this issue nicely by having the same behavior on X11 and
Wayland.

https://bugzilla.gnome.org/show_bug.cgi?id=758876