opensubtitles: Fix delete-event callback signature
gtk_widget_hide_on_delete (widget) is not a callback function. Using
it as a callback function to "widget::delete-event" signal ( which
takes 3 args for callback ) in C code works, though this is not
correct. Python checks for argument count of callbacks, and so this
hack doesn't work for python code. So, we move it UI file to trigger
native callback.

https://bugzilla.gnome.org/show_bug.cgi?id=775657