build: Renamed desktop file
desktop file has been renamed to make use of the '.in.in' extension
instead of the meson file, which was added to point the meson build
system, to be consistent with the rest of the files.

The '.in.in' extension has been chosen to denote two steps in its
generation, being the first one the variable substitution and the
second one the translation merge.

org.gnome.Totem.desktop.in.in.in has also been removed which is
unnecessary for the meson build port.

https://bugzilla.gnome.org/show_bug.cgi?id=784179