build: Add a dependency on generated libbacon files
In cases of large parallelism, the totem files may be compiled before
the libbacon headers are generated. The declare_dependency() object can
be instructed to transfer that dependency not just on the library to
link with, but also to the sources that need to exist in order to build
other targets that depend on those sources.