build: Improve debug handling
totem is considered a debug build if the build type is not
`release`. However this is wrong, because `plain` type should not
be considered a debug build.

The way compiler flags are handled on debug build has also been
changed by taking advantage of the `get_supported_arguments` helper
function. This function was added in 0.43, so meson's version has
been bumped.