build: Improve backend build
The backend build is creating different variables that are used only
once, so the common variables can be reused.

The `libbaconvideowidget` dependency also needs to reference the
created enum header and the `clutter-gtk` library so they both have
been added.

The compiler flags and dependencies of `bvw-test` have also been
revised.