plugins: Migrate plugins from Intltool to Gettext
This patch migrates plugins files from Intltool to Gettext by using
meson's i18n features. However, the migration has workarounds for
two issues.

The first issue lies in the fact that, even though the `plugin` uses
the Desktop entry specification[0], it is not recongnized as such by
gettext because it uses the file extension to detect the supported
files.

The second issue resides in the translatable entries used by the
`plugin` files: `Name` and `Description`. Although `Name` is part
of the default keyworks, `Description` is not.

As a workaround for these problems, the files has been renamed to
include the `desktop` extension, which helps gettext detecting
the `plugin` files. After the translation merge process they are
renamed to their right name without the `desktop` extension.

The default procedure to extract the translatable strings and
merge them has also been changed. New targets have been included
to generate the `pot` file and merge the translations.

[0] https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html