build: Check D-Bus session bus services directory
Totem install a session bus service file. This file is installed in
the `dbus-1/services` directory under `datadir`. However, the real
directory can be checked from the `dbus-1.pc` pkg-config file.

A new option has also been added to set as a different session bus
services directory, so the user can install the file in it.