opensubtitles: Fix a function signature
Commit 6afabc102b4ac50cbf4497847fceeaae1dfa07f5 changed the method
signature to please pylint:

-    def _show_dialog (self, params, _):
+    def _show_dialog (self, _):

This effectively makes the method take a single parameter instead of
two.

However the method is typically called with two arguments. (it is an
action callback)

This completely broke the OpenSubtitles plugin, which wasn't able to
show its dialog any more.

This commit makes the code ignore all parameters of the method, so
Pylint is still happy, but fixes the issue.