CI: Improve the build script and deliver bundles
Current implementation only build the master branch of the repository.
Even if the CI runs in a branch flatpak-builder will pull its own copy of the repository as specified in the manifest.

This replaces the flatpak job with the template from wiki page.

This also has the positive side-effect of generating flatpak bundles for each commit and delivering them with gitlab review apps.

For more info see:
https://gitlab.gnome.org/GNOME/Initiatives/wikis/DevOps-with-Flatpak#flatpak-ci