main: Add Mute action
This allows "Mute" to be exported as a D-Bus action.

Ubuntu uses this action as one of its Desktop Actions

The Desktop Actions were rejected from inclusion in GNOME years
ago because they are static instead of dynamic run-time actions
(so the actions like Play and Mute show even when the app is closed
and no video is playing). See https://bugzilla.gnome.org/671088