main: Remove app menu
Move items from the app menu to a menu button in the main window. This
new hamburger menu will show in the content navigation/selection view
and include the same items that were available in the app menu.

As we do not want 2 hamburger menus in the player view, duplicate the
preferences and keyboard shortcuts menu items in the player menu.

See:
https://gitlab.gnome.org/GNOME/Initiatives/wikis/App-Menu-Retirement

Closes: #265