backend: Return more info about subtitles/audio tracks
Instead of simply returning a list of language codes for subtitles and
audio tracks, return a struct contains both the language code and the
audio codec. We will be able to use this to differentiate tracks better
in the menus.

There are no functional or UI changes in this commit, just an API
extension.