main: Fix media player controls not appearing in gnome-shell
Media player controls in gnome-shell wouldn't appear if totem was first
launched normally, and a video selected afterwards, but would work if
one double-clicked on a file.

The problem was that the controls would only show if the CanPlay MPRIS
property was set to true, and the plugin would only change that property
if it received a property notification for an MRL being set. Nothing
was sending that property notification though.

One line later, the property is sent, and the MPRIS plugin works whether
or not totem was opened with a double-click.

Closes: #208