main: Update state of previous/next when chapters info arrives
This fixes the previous/next buttons not being sensitive when a single
file with chapter data is read as the "hey, a chapter is available now"
would come in later than we usually check.

The keyboard shortcuts for previous/next track already worked as whether
a chapter marker is available is checked when the key has been pressed.

Closes: #215