ci: flatpak: remove meson-args variable
The CI template was reworked to be able to parse the build
args from the manifest itself so we no longer need to
write them manually.

The template is no longer making use of the MESON_ARGS var.
If you need to override the args of the manifest use CONFIG_OPTS [2]

[1] https://gitlab.gnome.org/GNOME/citemplates/-/merge_requests/15
[2] https://gitlab.gnome.org/GNOME/citemplates/-/merge_requests/23#note_881242