Add 'Music' keyword to desktop file
totem has been GNOME's default music player for as long as I can
remember, since Music cannot open music files and Rhythmbox was never
part of GNOME core.

Now downstream is finally catching up by removing Rhythmbox:
https://pagure.io/fedora-workstation/issue/164

One concern is that users will have difficulty discovering that totem is
how you play music. Adding a desktop file keyword is not really a
spectacular way to address this, but it's better than nothing.

Design team is currently pursuing a strategy of trying to ensure users
understand that music should be played in Videos without redesiging or
rebranding the app. I'm not really sure if this is a great approach, but
we don't currently have consensus to do more than add the keyword.