preferences: Embed preferences.ui as a GResource
Instead of installing preferences.ui and opening the file, embed it
as a GResource and load it using gtk_builder_new_from_resource().
We don't need extra caution because the file is now guaranteed to
be embedded into the application.

This will be very handy for the next commit where we start using
a GtkBuilder template for the preferences dialog.