preferences: Add TotemPreferencesDialog
Most of the code is what was there in preferences.c, but with a twist: it now
has a GtkDialog subclass handling the preferences.

The laborious part of this change is replacing the old GtkBuilder-based
mechanism to retrieve widgets by their name with binding template children.
Signal callbacks can now be static, and be passed explicitly instead of using
G_MODULE_EXPORT to preserve their symbols.

This commit is the smallest possible way to achieve this, and doesn't make
use of various niceties and good practices of GTK development. Next commits
will cleanup and improve preferences to actually use more modern GTK practices.