preferences-dialog: Remove GtkHButtonBox
This button box has a single child, a GtkButton, and
is effectively only used to horizontally align the
button at the end. This property can be trivially set
in the button itself nowadays.

Remove the GtkHButtonBox holding the "Reset to Defaults"
button, and preserve the end alignment of the button by
setting the "halign" to "end" in the button itself.