playlist: Port to GTask
GSimpleAsyncResult is deprecated, and it makes compiling
Totem a bit noisy.

The port is pretty straightforward, since TotemPlaylist
usage of GSimpleAsyncResult is not complicated. The major
difference is that GTasks can only be returned once, so
replace the following (pseudo)code:

```
if (error)
  async_result.set_error(error)
async_result.set_boolean(result)
async_result.complete()

```

by this:

```
if (error)
  task.return_error(error)
else
  task.return_boolean(result)

```