backend: Remove GInitable interface
BaconVideoWidget could fail to initialise if GStreamer plugins were
missing, but GStreamer is initialised after options are parsed in the
GtkApplication's ->startup vfunc, which we want to use a GtkBuilder
template for, which has the fallible widget inside it.

This means that we can't be using a GInitable as we don't want to
have unreportable errors during the template creation[1].

Instead, move all the initialisation inside the GObject's _init()
function and add a function to check on its success and propagate an
error.

See https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/4461