backend: Move subtitles/languages list caching to video widget
Instead of having the front-end keep track of track lists, and compare
them, keep them inside the video widget. This also fixes the old
totem-menu code that didn't check for title equality when comparing
lists.

Note that this updates the subtitles and languages menu too often,
as we're *always* updating the menu even when it hasn't changed.

Fixes: becbe7f8ef61f51ccbbb228f248b85b02e1eeab9