backend: Fix "None" subtitle not selected when opening file
When opening a file, the GST_PLAY_FLAG_TEXT flag is set because we
want to be able to show a default subtitle track if any is selected.
But that means that bacon_video_widget_get_subtitle() will return
-1 (automatic) instead of -2 (none).

Check whether we have a single subtitle track (which must be "none") and
claim it was selected when checking the current subtitle.