grilo: Fix crash when scrolling for more content in some cases
When popping down the search entry, and scrolling down to activate the
"load content" action, we would try and load more search results but we
never really started a search, leading to a crash.

Make sure that a search is only considered to be started when the search
entry has been activated.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=2079657