main: Fix scroll-by-page GTK setting breaking slider
gtk-primary-button-warps-slider can be set to FALSE , a click will
cause the slider/value to move by the range’s page-size towards the
point clicked.

totem expects that a click will "warp" the slider to the location.

Fix by forcing this setting to TRUE for the GtkRange when the slider
is clicked.

Closes: #541