help: Prefer P as a play/pause shortcut
This avoids us having to explain what was on our old Wiki page:
Q: Why doesn't the space bar work to play/pause in windowed mode? It
only works in fullscreen, and sometimes in windowed mode.

A: The space bar will work in fullscreen. But when in windowed mode,
for accessibility reasons, we cannot use the space to play/pause (it
is used to "push" buttons already). It will work sometimes, as the
play/pause button, or the video widget might be the one with the
keyboard focus. Try using "P" to play/pause instead.