thumbnailer: Set number of threads to 1 for common decoders
Various decoders default to the number of CPU cores (or threads) for
decoder threads, requiring the allocation of output buffers for each
of them. On system with many cores this can cause the allocations to
exceed the RLIMIT_DATA memory limit.

Set the number of threads to 1. That should be all we need for the
thumbnailer use-case, preventing the mentioned cases and generally
reducing memory usage/allocations.

Unfortunately there isn't a generic way to do so, thus use
element-setup callback for well known codecs/properties for now.

See also
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/4423

Closes https://gitlab.gnome.org/GNOME/totem/-/issues/638